Time: 0 ms (100.00%), Space: 6 MB (62.65%) - LeetHub

+++ 0050-powx-n/0050-powx-n.cpp
@@ -21,7 +21,7 @@ public:
             }                
         }
         if(n<0)
-            ans = (double)1.0/(double)ans;
+            ans = 1.0/(double)ans;
         return ans;
     }
 };
